Move tarball extraction to its own source file
